% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscpstat.R
\name{dscpstat}
\alias{dscpstat}
\title{Descriptive Statistics}
\usage{
dscpstat(x)
}
\arguments{
\item{x}{The variable}
}
\value{
Descriptive statistics of x
}
\description{
Descriptive Statistics
}
\examples{
dscpstat(c(5:500))
}
