% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Calculate geometric mean with 95\% confidence intervals}
\usage{
geomean(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
geometric mean values and 95\% confidence intervals rounded to 2 decimal places
}
\description{
Calculate geometric mean with 95\% confidence intervals
}
