% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGPEstatsApp.R
\name{LGPEstatsApp}
\alias{LGPEstatsApp}
\title{Stat calculation app for LGPE}
\usage{
LGPEstatsApp(...)
}
\arguments{
\item{...}{arguments to be passed to other methods}
}
\description{
The purpose of the app is to help the runner plan ahead
using the information gained from the stat values at each level up.
Each stat is considered independent (the calculation does not take into
account the total AVs gained across all stats). At each level when stats
were seen by the runner, they can be adjusted in the app to narrow down
the IV-AV combination and stat range projection of each stat. Some
other useful information is shown, mainly to help the runner avoid
mistakes with inputs. Note that the app does not prevent invalid states
(e.g. a stat value that is impossible), it is up to the user to avoid that.
}
