% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltplm.R
\name{mltplm}
\alias{mltplm}
\title{Multiple Linear Regression with 2 predictor variables}
\usage{
mltplm(y, x1, x2)
}
\arguments{
\item{y}{The response variable}

\item{x1}{The first predictor variable}

\item{x2}{The second predictor variable}
}
\value{
b0, b1, b2
}
\description{
Multiple Linear Regression with 2 predictor variables
}
\examples{
mltplm(y<-(50:149),x1<-(1:100),x2<-(100:199))
}
