% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_calculator.R
\name{my_CI_upper}
\alias{my_CI_upper}
\title{Upper CI function}
\usage{
my_CI_upper(value, sd, prob = 0.95)
}
\arguments{
\item{value}{Input parameter for the function}

\item{sd}{Standard deviation of the sample}

\item{prob}{The probability for the confidence interval}
}
\value{
A numeric value for the upper CI
}
\description{
Upper CI function
}
\examples{
my_CI_upper(10,100)
}
