% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_calculator.R
\name{norm_ci}
\alias{norm_ci}
\title{CI function(Normal)}
\usage{
norm_ci(vect, conf.int = 0.95)
}
\arguments{
\item{vect}{Input parameter for the function}

\item{conf.int}{The desired confidence interval}
}
\value{
A numeric value for the lower CI
}
\description{
CI function(Normal)
}
\examples{
norm_ci(sample(1:10,20, replace=TRUE))
}
