% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplm.R
\name{simplm}
\alias{simplm}
\title{Simple linear regression with 1 predictor}
\usage{
simplm(x, y)
}
\arguments{
\item{x}{The predictor variable}

\item{y}{The response variable}
}
\value{
Scattterplot with fitted line, residual plot, b0, b1, r
}
\description{
Simple linear regression with 1 predictor
}
\examples{
simplm(x<-(1:100),y<-(50:149))
}
