% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statCalcFun.R
\name{statCalcFun}
\alias{statCalcFun}
\title{Stat calculation formula for LGPE}
\usage{
statCalcFun(
  pokemon = "eevee",
  stat = 1,
  iv = 31,
  av = 0,
  level = 5,
  nature = 1,
  happiness = 1
)
}
\arguments{
\item{pokemon}{the pokemon species (*_starter for starters)}

\item{stat}{stat id (1-6)}

\item{iv}{IV for the stat}

\item{av}{number of AVs gained in the stat}

\item{level}{level of the pokemon}

\item{nature}{nature id (1-25); see nature.names}

\item{happiness}{happiness modifier (1 + (1-10)\%)}
}
\description{
Stat calculation formula for LGPE
}
